stock-dashboard/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
yfinance==0.2.40
websockets==12.0
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
praw==7.7.1
newsapi-python==0.2.7
transformers==4.41.2
torch==2.3.1
scikit-learn==1.5.0
joblib==1.4.2
redis==5.0.4
python-dotenv==1.0.1
httpx==0.27.0
ta==0.11.0
requests==2.32.3
